refactor: moved invoice components to subfolder
