+++ apps/dashboard/src/modules/financial/components/invoice/InvoiceInfo.vue
