+++ apps/dashboard/src/modules/financial/components/invoice/InvoiceStepsCard.vue
