+++ apps/dashboard/src/modules/financial/components/invoice/InvoiceTable.vue
