+++ apps/dashboard/src/modules/financial/components/invoice/forms/InvoiceEntriesTable.vue
